refactor(ui): keep motion gate files inside the shared line budgets
